+++ tests/Feature/Framework/Plugins/Module/Infrastructure/ExamplePluginWithConstructor.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace GacelaTest\Feature\Framework\Plugins\Module\Infrastructure;
 
 use Gacela\Framework\Container\Container;
-use Gacela\Framework\Plugin\PluginInterface;
+
 use GacelaTest\Fixtures\StringValue;
 
-final class ExamplePluginWithConstructor implements PluginInterface
+final class ExamplePluginWithConstructor
 {
     public function __construct(
         private Container $container,